FEATURE: replace the naive "fast Top-N" algorithm with TPUT
TPUT is an algorithm to quickly answer Top-N queries (e.g., find the N objects
with the highest aggregate values) in a distributed network. It replaces a
similar algorithm called "fast Top-N", mostly because it was not proven that it
works correctly. TPUT is, because it is a modified TPUT algorithm presented in
"Efficient Top-K Query Calculation in Distributed Networks" by Pei Cao and Zhe
Wang.
It supports both descending and ascending order directions.